pyproject.toml
[tool.poetry]
name = "ATS"
version = "0.1.0"
description = "Applicant Tracking System"
authors = ["Vladyslav Luchka <luchka.vl@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.85.0"}
SQLAlchemy = "^1.4.41"
psycopg2 = "^2.9.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
